Revert "Added normalization with percentage"
This reverts commit c48e3e475b1da90f4163557629721f01ef76d4be.

+++ Algorithm/AlgoritmProtocol.h
@@ -16,7 +16,7 @@ typedef struct
 typedef struct
 {
     int id;
-    double recommendation[96];
+    int recommendation[96];
     time_start time[96];
 } AlgoritmResult;
